+++ pages/_documents.js
@@ -28,20 +28,3 @@ export default class MyDocument extends Document {
 		);
 	}
 }
-
-/*export default function Document() {
-	return (
-		<Html>
-			<Head>
-				<link
-					rel="stylesheet"
-					href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap"
-				/>
-			</Head>
-			<body>
-				<Main />
-				<NextScript />
-			</body>
-		</Html>
-	);
-}*/
