+++ src/components/Button/StyledButton.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+const EmailButton = styled.button`
+	text-transform: uppercase;
+	color: white;
+	background-color: #db6c4c;
+	border-radius: 19px;
+	border: none;
+	display: flex;
+	justify-content: center;
+	padding: 7px 13px;
+`;
+
+export default EmailButton;
